some fixes

+++ app/api/(admin)/product-attribute-value/[id]/route.ts
@@ -1,12 +1,32 @@
 import { NextResponse } from "next/server";
 import { Prisma } from "@/lib/prisma";
+import { allowedRoleForCategoryAndProductManagement } from "@/lib/constants";
+import { auth } from "@/lib/auth";
+import serverResponse from "@/lib/serverResponse";
+import { ROLE } from "@prisma/client";
 
 export async function GET(
     request: Request,
     { params }: { params: Promise<{ id: string }> },
 ) {
     try {
-        // TODO: AUTHENTICATION
+        const session = await auth();
+        if (
+            !session ||
+            session.user.userType != "staff" ||
+            !allowedRoleForCategoryAndProductManagement.includes(
+                session.user.staff?.role as ROLE,
+            ) ||
+            (session.user.staff?.role !== "ADMIN" &&
+                session.user.staff?.isBanned)
+        ) {
+            return serverResponse({
+                status: 401,
+                success: false,
+                message: "Unauthorized",
+            });
+        }
+
         const { id } = await params;
 
         const productAttributeValue =
@@ -39,9 +59,44 @@ export async function DELETE(
     { params }: { params: Promise<{ id: string }> },
 ) {
     try {
-        // TODO: AUTHENTICATION
+        const session = await auth();
+        if (
+            !session ||
+            session.user.userType != "staff" ||
+            !allowedRoleForCategoryAndProductManagement.includes(
+                session.user.staff?.role as ROLE,
+            ) ||
+            (session.user.staff?.role !== "ADMIN" &&
+                session.user.staff?.isBanned)
+        ) {
+            return serverResponse({
+                status: 401,
+                success: false,
+                message: "Unauthorized",
+            });
+        }
         const { id } = await params;
-
+        const value = await Prisma.productAttributeValue.findUnique({
+            where: {
+                id: parseInt(id),
+            },
+            include: {
+                _count: {
+                    select: {
+                        productItems: true,
+                    },
+                },
+            },
+        });
+        if (!!value?._count?.productItems && value?._count?.productItems > 0) {
+            return NextResponse.json(
+                {
+                    success: false,
+                    message: "product attribute value can't delete.",
+                },
+                { status: 400 },
+            );
+        }
         const productAttributeValue = await Prisma.productAttributeValue.delete(
             {
                 where: { id: parseInt(id) },
